Build-less `cheatsheet` template
